美化多个源代码文件中的函数名和变量名，提升代码可读性
- 99_unmatched_functions.c: 重命名FUN_1808a3e1c为ProcessSystemDataStreamAndValidation，添加详细文档注释
- 其他文件: 自动优化变量命名和函数结构

+++ TaleWorlds.Native/src/01_initialization.c
@@ -8063,12 +8063,12 @@ void InitializeSystemResourceInitializer(void)
 void InitializeSystemFloatingPointCalculator(void)
 
 {
-  ulong long TableIndex;
+  uint64_t TableIndex;
   float *TableEntryPointer;
   int NodeIdentifierComparisonResult;
-  ulong long OuterLoopCounter;
-  uint InnerLoopCounter;
-  ulong long BaseCounter;
+  uint64_t OuterLoopCounter;
+  uint32_t InnerLoopCounter;
+  uint64_t BaseCounter;
   int RangeOffset;
   float *CurrentTableAddress;
   float CalculatedValue;
@@ -8130,8 +8130,8 @@ void InitializeSystemEventManager(void)
   char NodeActiveFlag;
   void** SystemDataTable;
   int SystemIdentifierCompareResult;
-  long long* MemorySystemDataPointer;
-  long long SystemCurrentOperationTimestamp;
+  int64_t* MemorySystemDataPointer;
+  int64_t SystemCurrentOperationTimestamp;
   void** RootNodeReference;
   void** CurrentNodePointer;
   void** NextNodePointer;
@@ -8189,8 +8189,8 @@ void InitializeSystemSearchManager(void)
   char NodeActiveFlag;
   void** SystemDataTable;
   int SystemIdentifierCompareResult;
-  long long* MemorySystemDataPointer;
-  long long SystemCurrentOperationTimestamp;
+  int64_t* MemorySystemDataPointer;
+  int64_t SystemCurrentOperationTimestamp;
   void** RootNodeReference;
   void** CurrentNodePointer;
   void** NextNodePointer;
@@ -8247,8 +8247,8 @@ void InitializeSystemIndexManager(void)
   char NodeActiveFlag;
   void** SystemDataTable;
   int SystemIdentifierCompareResult;
-  long long* MemorySystemDataPointer;
-  long long SystemCurrentOperationTimestamp;
+  int64_t* MemorySystemDataPointer;
+  int64_t SystemCurrentOperationTimestamp;
   void** RootNodeReference;
   void** CurrentNodePointer;
   void** NextNodePointer;
